# frntEnd/openapi.yaml
openapi: 3.1.0
info:
  title: AI Workflow Builder API
  version: 1.0.0
  description: |
    REST API for the AI Workflow Builder platform.

    ## Authentication
    All endpoints (except `/auth/*` and `/health`) require a valid JWT Bearer token
    or an API Key in the `X-API-Key` header.

    ## Tenant Isolation
    Every request is scoped to a tenant. The tenant is resolved from the JWT claim
    (`tenant_id`) or from the API key row. Cross-tenant access is impossible by design.

    ## Error Envelope
    All error responses use the standard envelope below:
    ```json
    {
      "error": {
        "code": "WORKFLOW_NOT_FOUND",
        "message": "Workflow wf_abc123 not found.",
        "request_id": "req_01J..."
      }
    }
    ```

  contact:
    name: Platform Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.workflowplatform.io/v1
    description: Production (us-east-1)
  - url: https://api.eu.workflowplatform.io/v1
    description: EU Production (eu-west-1 — v1.2+)
  - url: http://localhost:8000/v1
    description: Local development

# ─────────────────────────────────────────────────────────────────────────────
# Security Schemes
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256-signed JWT. Obtain via `/auth/login` or `/auth/token/refresh`.

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Tenant API key (prefix `wfk_`). Created in the dashboard.
        Scopes are checked per-endpoint.

  # ───────────────────────────────────────────────────────────────────────────
  # Reusable Schemas
  # ───────────────────────────────────────────────────────────────────────────
  schemas:
...
    ErrorDetail:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: WORKFLOW_NOT_FOUND
        message:
          type: string
          example: "Workflow wf_abc123 not found."
        request_id:
          type: string
          example: req_01JRXYZ
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'

    # ── Pagination ─────────────────────────────────────────────────────────────
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        has_next:
          type: boolean

    # ── Auth ───────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, tenant_id, email, role, is_verified, mfa_enabled, created_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [OWNER, EDITOR, VIEWER]
        is_verified:
          type: boolean
        mfa_enabled:
          type: boolean
        sso_provider:
          type: string
          nullable: true
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time

    # ── API Keys ───────────────────────────────────────────────────────────────
    ApiKeyCreate:
      type: object
      required: [name, scopes]
      properties:
        name:
          type: string
          maxLength: 100
        scopes:
          type: array
          items:
            type: string
            enum:
              - workflows:read
              - workflows:write
              - executions:read
              - executions:write
              - executions:trigger
              - logs:read
        expires_at:
          type: string
          format: date-time
          nullable: true
    ApiKeyCreated:
      type: object
      description: Full key is returned ONCE at creation time — never retrievable again.
      required: [id, name, key, key_prefix, scopes, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key:
          type: string
          description: Full API key — store securely, shown only once.
          example: wfk_01JRXYZ...
        key_prefix:
          type: string
          example: wfk_01JR
        scopes:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    ApiKeyInfo:
      type: object
      description: Safe metadata — never exposes the full key.
      required: [id, name, key_prefix, scopes, is_active, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key_prefix:
          type: string
        scopes:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        expires_at:
          type: string
          format: date-time
          nullable: true
        last_used_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time

    # ── Workflows ──────────────────────────────────────────────────────────────
...
    TriggerRequest:
      type: object
      properties:
        input:
          type: object
          description: Trigger payload. Shape validated against workflow's trigger schema.
        version_no:
          type: integer
          nullable: true
          description: Pin to a specific version. Omit to use the latest published version.
    ExecutionRun:
      type: object
      required: [run_id, tenant_id, workflow_id, status, started_at]
      properties:
        run_id:
          type: string
        tenant_id:
          type: string
          format: uuid
        workflow_id:
          type: string
        version_no:
          type: integer
        status:
          type: string
          enum: [QUEUED, RUNNING, SUCCESS, FAILED, CANCELLED, WAITING_HUMAN]
        trigger_input:
          type: object
          description: Original trigger payload (may be redacted by PII policy).
        output:
          type: object
          nullable: true
          description: Final workflow output. Null if not yet complete.
        error:
          type: object
          nullable: true
          properties:
            node_id:
              type: string
            message:
              type: string
            traceback:
              type: string
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
        retention_expires_at:
          type: string
          format: date-time
          nullable: true
    NodeExecutionState:
      type: object
      required: [node_id, run_id, status]
      properties:
        node_id:
          type: string
        run_id:
          type: string
        status:
          type: string
          enum: [PENDING, RUNNING, SUCCESS, FAILED, SKIPPED, WAITING_HUMAN]
        input:
          type: object
          nullable: true
        output:
          type: object
          nullable: true
        error_message:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        ended_at:
          type: string
          format: date-time
          nullable: true
        duration_ms:
          type: integer
          nullable: true
        isolation_tier:
          type: integer
          minimum: 0
          maximum: 3
    HumanInputSubmit:
      type: object
      required: [run_id, node_id, response]
      properties:
        run_id:
          type: string
        node_id:
          type: string
        response:
          type: object
          description: Free-form response matching the node's expected schema.
        approved:
          type: boolean
          description: For approval gates — true to proceed, false to reject.

    # ── Execution Logs ─────────────────────────────────────────────────────────
    LogEntry:
      type: object
      required: [timestamp, level, node_id, message]
      properties:
        timestamp:
          type: string
          format: date-time
        level:
          type: string
          enum: [DEBUG, INFO, WARNING, ERROR]
        run_id:
          type: string
        node_id:
          type: string
        message:
          type: string
        metadata:
          type: object
          nullable: true

    # ── Schedules ──────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: UNAUTHORIZED
              message: "Bearer token is missing or expired."
              request_id: req_01JRXYZ
    Forbidden:
      description: Authenticated but insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found (or not visible to this tenant).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Resource state conflict (e.g. duplicate slug, optimistic lock).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    UnprocessableEntity:
      description: Validation error — request body failed schema or business rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    TooManyRequests:
      description: Rate limit or quota exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the limit resets.
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'

  # ───────────────────────────────────────────────────────────────────────────
  # Reusable Parameters
  # ───────────────────────────────────────────────────────────────────────────
  parameters:
    WorkflowId:
      name: workflow_id
      in: path
      required: true
      schema:
        type: string
      description: Workflow identifier (e.g. wf_01JRXYZ).
    RunId:
      name: run_id
      in: path
      required: true
      schema:
        type: string
...
    KeyId:
      name: key_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

# ─────────────────────────────────────────────────────────────────────────────
# Global Security (overridable per-operation)
# ─────────────────────────────────────────────────────────────────────────────
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Tags
# ─────────────────────────────────────────────────────────────────────────────
...
paths:
...
  /users/me:
    get:
      operationId: users_get_me
      tags: [Users]
      summary: Get current user profile
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      operationId: users_update_me
      tags: [Users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                current_password:
                  type: string
                  format: password
                  description: Required when changing password or email.
                new_password:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Profile updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/api-keys:
    get:
      operationId: users_list_api_keys
      tags: [Users]
      summary: List API keys for current user
      responses:
        '200':
          description: API key list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKeyInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: users_create_api_key
      tags: [Users]
      summary: Create a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
      responses:
        '201':
          description: API key created. Full key shown only once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreated'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/api-keys/{key_id}:
    delete:
      operationId: users_revoke_api_key
      tags: [Users]
      summary: Revoke an API key
      parameters:
        - $ref: '#/components/parameters/KeyId'
      responses:
        '204':
          description: Key revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Workflows ─────────────────────────────────────────────────────────────────
...
  /workflows/{workflow_id}/trigger:
    post:
      operationId: executions_trigger
      tags: [Executions]
      summary: Trigger a workflow execution (async — returns run_id immediately)
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerRequest'
      responses:
        '202':
          description: Execution accepted and queued.
          content:
            application/json:
              schema:
                type: object
                required: [run_id, status]
                properties:
                  run_id:
                    type: string
                  status:
                    type: string
                    enum: [QUEUED]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /executions:
    get:
      operationId: executions_list
      tags: [Executions]
      summary: List execution runs for the current tenant
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: workflow_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [QUEUED, RUNNING, SUCCESS, FAILED, CANCELLED, WAITING_HUMAN]
        - name: started_after
          in: query
          schema:
            type: string
            format: date-time
        - name: started_before
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Paginated execution list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExecutionRun'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /executions/{run_id}:
    get:
      operationId: executions_get
      tags: [Executions]
      summary: Get execution run status and output
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Execution run detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionRun'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /executions/{run_id}/cancel:
    post:
      operationId: executions_cancel
      tags: [Executions]
      summary: Cancel a running or queued execution
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Cancellation accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionRun'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
          description: Run is already in a terminal state.
  /executions/{run_id}/retry:
    post:
      operationId: executions_retry
      tags: [Executions]
      summary: Retry a failed execution (creates a new run)
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '202':
          description: New retry execution queued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                  status:
                    type: string
                    enum: [QUEUED]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  # ── Node Executions ───────────────────────────────────────────────────────────
  /executions/{run_id}/nodes:
    get:
      operationId: nodes_list
      tags: [Nodes]
      summary: List all node execution states for a run
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Node states for the run.
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/NodeExecutionState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /executions/human-input:
    post:
      operationId: executions_human_input
      tags: [Nodes]
      summary: Submit human-input response to resume a waiting execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HumanInputSubmit'
      responses:
        '200':
          description: Input accepted. Execution resumed (or rejected).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionRun'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  # ── Logs ──────────────────────────────────────────────────────────────────────
  /executions/{run_id}/logs:
    get:
      operationId: logs_list
      tags: [Logs]
      summary: Fetch structured logs for an execution run
      parameters:
        - $ref: '#/components/parameters/RunId'
        - name: node_id
          in: query
          schema:
            type: string
          description: Filter logs to a specific node.
        - name: level
          in: query
          schema:
            type: string
            enum: [DEBUG, INFO, WARNING, ERROR]
        - name: since
          in: query
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Log entries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/LogEntry'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Schedules ─────────────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: team_list_members
      tags: [Team]
      summary: List all team members in the current tenant
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Team member list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...